pyproject.toml
[tool.poetry]
name = "camerahub"
version = "0.1.0"
description = "App for cataloguing vintage cameras, lenses, films, negatives & prints"
authors = ["Jonathan Gazeley <me@jonathangazeley.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4"
django-material = "^1.10.0"
django-countries = "^7.3.2"
django-choices = "^1.7.2"
django-colorfield = "^0.7.0"
numpy = "^1.22.4"
django-geoposition-2 = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
autopep8 = "^1.6.0"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
